IM.Base +Se creo la carpeta Converters: la idea es ir agregando todos los converters que utilizamos en el proyecto a esta carpeta
ya se movieron
+ByteToBooleanValueConverter
+MathOperationConverter

+++ IM.Base/Classes/InvitationValidationRules.cs
@@ -1082,8 +1082,8 @@ namespace IM.Base.Classes
     {
       if (!string.IsNullOrWhiteSpace(number))
       {
-        int numb = 0;
-        int.TryParse(number, out numb);
+        decimal numb = 0;
+        decimal.TryParse(number, out numb);
         if (blnValidateBounds)
         {
           if (numb > maxNumber)
